requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==38.0.4
Deprecated==1.2.13
dill==0.3.6
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.87.0
frozenlist==1.4.0
googletrans==4.0.0rc1
greenlet==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.12.0
kombu==5.2.4
lazy-object-proxy==1.9.0
lxml==4.9.3
mccabe==0.7.0
multidict==6.0.4
openai==0.28.1
packaging==21.3
passlib==1.7.4
platformdirs==3.5.3
prompt-toolkit==3.0.32
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pylint==2.17.4
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.6
redis==4.3.4
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
starlette==0.21.0
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==2.0.6
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.9.2
